+++ src/main/java/nl/saxion/refactor/model/Printer.java
@@ -3,7 +3,7 @@ package nl.saxion.refactor.model;
 import nl.saxion.refactor.model.visitor.PrinterVisitor;
 
 public interface Printer {
-    int getId();
+    Long getId();
     String getName();
     boolean printFits(Print print);
     int calculatePrintTime(String filename);
